update weather forecast view

+++ lib/data/api/endpoints/weather_endpoints.dart
@@ -1,5 +1,5 @@
 abstract class WeatherEndpoints {
   static const String forecast = 'forecast?';
-  static String byLatAndLon(int lat, int lon) => '${forecast}lat=$lat&lon=$lon';
+  static String byLatAndLon(double lat, double lon) => '${forecast}lat=$lat&lon=$lon';
   static String byCityName(String cityName) => '${forecast}q=$cityName';
 }
